🔥 chore: move src/api/controllers directory to src/controllers
