+++ src/routes/users.ts
@@ -2,8 +2,8 @@ import { Router } from "express";
 
 import Joi from "joi";
 
-import verifyToken from "../api/middlewares/verifyToken";
-import validateSchema from "../api/middlewares/validateSchema";
+import verifyToken from "../middlewares/verifyToken";
+import validateSchema from "../middlewares/validateSchema";
 
 import * as UserController from "../controllers/user";
